fix: Codegens gifts. Unifies staff names
